tests: split off test_file_items.py
